# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Product {
  id: number;
  nameEn: string;
  nameAr: string;
  descriptionEn: string;
  descriptionAr: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  subcategory?: string | null;
  images: string[];
  videoUrl?: string | null;
  sizes: string[];
  colors: string[];
  fabricEn?: string | null;
  fabricAr?: string | null;
  inStock: boolean;
  featured: boolean;
  badge?: string | null;
  rating: number;
  reviewCount: number;
  createdAt?: string | null;
}

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  quantity: number;
  size: string;
  color: string;
  product?: Product;
}

export interface AddToCartRequest {
  productId: number;
  size: string;
  color: string;
  quantity?: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  nameEn: string;
  nameAr: string;
  price: number;
  quantity: number;
  size: string;
  color: string;
}

export interface Order {
  id: number;
  sessionId: string;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  discount: number;
  total: number;
  status: string;
  fulfillmentType: string;
  pickupStoreId?: number | null;
  pickupStoreName?: string | null;
  pickupAddress?: string | null;
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  shippingCountry: string;
  discountCode?: string | null;
  notes?: string | null;
  createdAt?: string | null;
}

export interface CreateOrderRequest {
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  discount: number;
  total: number;
  fulfillmentType: string;
  pickupStoreId?: number | null;
  pickupStoreName?: string | null;
  pickupAddress?: string | null;
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  shippingCountry: string;
  discountCode?: string | null;
  notes?: string | null;
}

export interface Store {
  id: number;
  nameEn: string;
  nameAr: string;
  addressEn: string;
  addressAr: string;
  city: string;
  phone: string;
  hoursEn: string;
  hoursAr: string;
  mapUrl?: string | null;
  isActive: boolean;
}

export interface ValidateDiscountRequest {
  code: string;
}

export interface DiscountValidation {
  type: string;
  value: number;
  minOrder?: number | null;
}

export type GetProductsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
};