# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Masah Boutique API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: stores
    description: Store operations
  - name: discounts
    description: Discount operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Added cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get session orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stores:
    get:
      operationId: getStores
      tags: [stores]
      summary: Get active stores
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"

  /discounts/validate:
    post:
      operationId: validateDiscount
      tags: [discounts]
      summary: Validate a discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateDiscountRequest"
      responses:
        "200":
          description: Discount info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountValidation"
        "404":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Product:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        price:
          type: integer
        originalPrice:
          type: integer
          nullable: true
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        fabricEn:
          type: string
          nullable: true
        fabricAr:
          type: string
          nullable: true
        inStock:
          type: boolean
        featured:
          type: boolean
        badge:
          type: string
          nullable: true
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - nameEn
        - nameAr
        - descriptionEn
        - descriptionAr
        - price
        - category
        - images
        - sizes
        - colors
        - inStock
        - featured
        - rating
        - reviewCount

    CartItem:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - sessionId
        - productId
        - quantity
        - size
        - color

    AddToCartRequest:
      type: object
      properties:
        productId:
          type: integer
        size:
          type: string
        color:
          type: string
        quantity:
          type: integer
      required:
        - productId
        - size
        - color

    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        price:
          type: integer
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
      required:
        - productId
        - nameEn
        - nameAr
        - price
        - quantity
        - size
        - color

    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: integer
        shipping:
          type: integer
        discount:
          type: integer
        total:
          type: integer
        status:
          type: string
        fulfillmentType:
          type: string
        pickupStoreId:
          type: integer
          nullable: true
        pickupStoreName:
          type: string
          nullable: true
        pickupAddress:
          type: string
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
          nullable: true
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingCountry:
          type: string
        discountCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - sessionId
        - items
        - subtotal
        - shipping
        - discount
        - total
        - status
        - fulfillmentType
        - customerName
        - customerPhone
        - shippingAddress
        - shippingCity
        - shippingCountry

    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: integer
        shipping:
          type: integer
        discount:
          type: integer
        total:
          type: integer
        fulfillmentType:
          type: string
        pickupStoreId:
          type: integer
          nullable: true
        pickupStoreName:
          type: string
          nullable: true
        pickupAddress:
          type: string
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
          nullable: true
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingCountry:
          type: string
        discountCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - items
        - subtotal
        - shipping
        - discount
        - total
        - fulfillmentType
        - customerName
        - customerPhone
        - shippingAddress
        - shippingCity
        - shippingCountry

    Store:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        addressEn:
          type: string
        addressAr:
          type: string
        city:
          type: string
        phone:
          type: string
        hoursEn:
          type: string
        hoursAr:
          type: string
        mapUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - nameEn
        - nameAr
        - addressEn
        - addressAr
        - city
        - phone
        - hoursEn
        - hoursAr
        - isActive

    ValidateDiscountRequest:
      type: object
      properties:
        code:
          type: string
      required:
        - code

    DiscountValidation:
      type: object
      properties:
        type:
          type: string
        value:
          type: integer
        minOrder:
          type: integer
          nullable: true
      required:
        - type
        - value