# lib/api-zod/src/generated/types/addToCartRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddToCartRequest {
  productId: number;
  size: string;
  color: string;
  quantity?: number;
}