# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface CreateOrderRequest {
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  discount: number;
  total: number;
  fulfillmentType: string;
  pickupStoreId?: number | null;
  pickupStoreName?: string | null;
  pickupAddress?: string | null;
  customerName: string;
  customerEmail?: string | null;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  shippingCountry: string;
  discountCode?: string | null;
  notes?: string | null;
}