# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  nameEn: string;
  nameAr: string;
  price: number;
  quantity: number;
  size: string;
  color: string;
}