# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  nameEn: string;
  nameAr: string;
  descriptionEn: string;
  descriptionAr: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  subcategory?: string | null;
  images: string[];
  videoUrl?: string | null;
  sizes: string[];
  colors: string[];
  fabricEn?: string | null;
  fabricAr?: string | null;
  inStock: boolean;
  featured: boolean;
  badge?: string | null;
  rating: number;
  reviewCount: number;
  createdAt?: string | null;
}