# lib/api-zod/src/generated/types/store.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masah Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Store {
  id: number;
  nameEn: string;
  nameAr: string;
  addressEn: string;
  addressAr: string;
  city: string;
  phone: string;
  hoursEn: string;
  hoursAr: string;
  mapUrl?: string | null;
  isActive: boolean;
}